scholarship form rename and refactor (#596)
